feat: PR conversation panel default reviewers
